{Cliente} Organização e implementação do itens do inventório
