+++ lib/xclarity_client.rb
@@ -8,3 +8,4 @@ require 'xclarity_client/client'
 require 'xclarity_client/xclarity_base'
 require 'xclarity_client/virtual_appliance_management'
 require 'xclarity_client/node'
+require 'xclarity_client/node_controller'
