+++ lib/xclarity_client/client.rb
@@ -6,11 +6,19 @@ module XClarityClient
     end
 
     def discover_nodes
-      Node.new(@connection).population
+      NodeController.new(@connection).population
     end
 
     def fetch_nodes(uuids)
-      array_nodes = Node.new(@connection).get_object_nodes(uuids)
+      array_nodes = NodeController.new(@connection).get_object_nodes(uuids)
+    end
+
+    def fetch_nodes_exclude_attr(uuids, attributes)
+      array_nodes = NodeController.new(@connection).get_object_nodes_exclude_attributes(uuids, attributes)
+    end
+
+    def fetch_nodes_include_attr(uuids, attributes)
+      array_nodes = NodeController.new(@connection).get_object_nodes_include_attributes(uuids, attributes)
     end
 
   end
